+++ src/DoctrineQueryHelper.php
@@ -18,25 +18,51 @@ use XApi\Repository\Doctrine\Mapping\Verb;
 /**
  * @author Mathieu Boldo <mathieu.boldo@entrili.com>
  */
-abstract class AvoidDuplicatesHelper
+abstract class DoctrineQueryHelper
 {
-    public static function findActor(QueryBuilder $queryBuilder, ?StatementObject $statementObject): StatementObject
+    public static function findActor(QueryBuilder $queryBuilder, ?StatementObject $statementObject): ?StatementObject
     {
         if (!$statementObject instanceof StatementObject) {
+            return $statementObject;
+        }
+
+        if ($statementObject->type !== StatementObject::TYPE_AGENT) {
             return $statementObject;
         }
 
         $queryBuilder
             ->select('o')
             ->from(StatementObject::class, 'o')
-            ->andWhere($queryBuilder->expr()->like('o.type', ':type'))
-            ->andWhere($queryBuilder->expr()->like('o.accountName', ':accountName'))
-            ->andWhere($queryBuilder->expr()->like('o.accountHomePage', ':accountHomePage'))
-            ->andWhere($queryBuilder->expr()->like('o.name', ':name'))
-            ->setParameter('type', 'agent')
-            ->setParameter('accountName', $statementObject->accountName)
-            ->setParameter('accountHomePage', $statementObject->accountHomePage)
-            ->setParameter('name', $statementObject->name);
+            ->andWhere($queryBuilder->expr()->like('o.type', ':objectType'))
+            ->setParameter('objectType', StatementObject::TYPE_AGENT);
+
+        $andX = $queryBuilder->expr()->andX();
+        $parameters = [];
+
+        if ($statementObject->accountName) {
+            $andX->add($queryBuilder->expr()->like('o.accountName', ':accountName'));
+            $parameters['accountName'] = $statementObject->accountName;
+        }
+
+        if ($statementObject->accountHomePage) {
+            $andX->add($queryBuilder->expr()->like('o.accountHomePage', ':accountHomePage'));
+            $parameters['accountHomePage'] = $statementObject->accountHomePage;
+        }
+
+        if ($statementObject->name) {
+            $andX->add($queryBuilder->expr()->like('o.name', ':name'));
+            $parameters['name'] = $statementObject->name;
+        }
+
+        if ($andX->count() === 0) {
+            return $statementObject;
+        }
+
+        $queryBuilder->andWhere($andX);
+
+        foreach ($parameters as $key => $value) {
+            $queryBuilder->setParameter($key, $value);
+        }
 
         if ($actors = $queryBuilder->getQuery()->getResult()) {
             return $actors[0]; // Link with first found
@@ -51,11 +77,27 @@ abstract class AvoidDuplicatesHelper
             return $statementObject;
         }
 
+        if ($statementObject->type !== StatementObject::TYPE_ACTIVITY) {
+            return $statementObject;
+        }
+
         $queryBuilder
             ->select('o')
             ->from(StatementObject::class, 'o')
             ->andWhere($queryBuilder->expr()->eq('o.activityId', ':activityId'))
             ->setParameter('activityId', $statementObject->activityId);
+
+        if ($statementObject->hasActivityName) {
+            $queryBuilder
+                ->andWhere($queryBuilder->expr()->eq('o.hasActivityName', ':hasActivityName'))
+                ->setParameter('hasActivityName', $statementObject->hasActivityName);
+        }
+
+        if ($statementObject->activityType) {
+            $queryBuilder
+                ->andWhere($queryBuilder->expr()->eq('o.activityType', ':activityType'))
+                ->setParameter('activityType', $statementObject->activityType);
+        }
 
         if ($activities = $queryBuilder->getQuery()->getResult()) {
             return $activities[0]; // Link with first found
@@ -75,6 +117,24 @@ abstract class AvoidDuplicatesHelper
             ->from(Context::class, 'c')
             ->where($queryBuilder->expr()->eq('c.registration', ':id'))
             ->setParameter('id', $context->registration);
+
+        if ($context->revision) {
+            $queryBuilder
+                ->andWhere($queryBuilder->expr()->eq('c.revision', ':revision'))
+                ->setParameter('revision', $context->revision);
+        }
+
+        if ($context->language) {
+            $queryBuilder
+                ->andWhere($queryBuilder->expr()->eq('c.language', ':language'))
+                ->setParameter('language', $context->language);
+        }
+
+        if ($context->platform) {
+            $queryBuilder
+                ->andWhere($queryBuilder->expr()->eq('c.platform', ':platform'))
+                ->setParameter('platform', $context->platform);
+        }
 
         if ($contexts = $queryBuilder->getQuery()->getResult()) {
             return $contexts[0];