+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 CHANGELOG
 =========
 
-0.2.0
+0.4
 -----
 
-* dropped suppport for PHP < 8.1
+* dropped support for PHP < 8.4
+* added unit and functional `StateRepositoryTest`
+
+0.3
+-----
+
+* dropped support for `doctrine/orm` < 3.0
+
+0.2
+-----
+
+* dropped support for PHP < 8.1
 * add QuoteStrategy
 * Replace "json_array" by "json" into metadata 
 * All dependencies from php-xapi/* are now loaded from forks at `https://github.com/evolution-job/`